Initial import of frequentist inference algorithms.
This is roughly what I used for the NIPS 2015 submission. It needs to be
reworked in order to fit in this library, so this should be considered
work-in-progress.

Furthermore, these algorithms are only for the logit / Luce model, I need to
provide some basic inference methods for the probit / Thurstone model.