Changed the convention w.r.t. rankings.
From now on, rankings induced by parameters are assumed to be as follows: the
item with lowest parameter is assigned rank 1, and the one with highest
parameter is assigned rank `n_items`.

This is simpler, because it is consistent with the natural order on the
parameters. But it is slightly unintuitive in certain cases, because the item
at rank 1 is the "worst" item, not the "best".